Separate execution provider from image inference
