+++ com.github.asus4.onnxruntime.unity/Runtime/ExecutionProviderOptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3679332aa952e42fda6e1c0210ff32e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
